Add run_pod_sandbox integration test
